Revert "Revert "Merge branch '2.0.0' into 1.0.0""
This reverts commit 02c8ed1493f9a58f5867740da7af84db689def3a.
