+++ constants/organization-constants.js
@@ -1,0 +1,8 @@
+module.exports = Object.freeze({
+    ORGANIZATION: {
+        STATUSES: {
+            ACTIVE: "Active",
+            INACTIVE: "Inactive",
+        },
+    },
+});
