+++ services/organization.js
@@ -1,3 +1,6 @@
+const {ERROR} = require("../constants/error-constants");
+const {USER} = require("../constants/user-constants");
+
 class Organization {
   constructor(organizationCollection) {
       this.organizationCollection = organizationCollection;
@@ -10,7 +13,44 @@ class Organization {
       return result?.length > 0 ? result[0] : null;
   }
 
-  async listOrganizations(filters) {
+  /**
+   * List Organizations API Interface.
+   *
+   * - `ADMIN` and `ORG_OWNER can call this API
+   * - `ORG_OWNER` is limited to only their organization
+   *
+   * @api
+   * @param {Object} params Endpoint parameters
+   * @param {{ cookie: Object, userInfo: Object }} context request context
+   * @returns {Promise<Object[]>} An array of Organizations
+   */
+  async listOrganizationsAPI(params, context) {
+    if (!context?.userInfo?.email || !context?.userInfo?.IDP) {
+        throw new Error(ERROR.NOT_LOGGED_IN)
+    }
+    if (context?.userInfo?.role !== USER.ROLES.ADMIN && context?.userInfo.role !== USER.ROLES.ORG_OWNER) {
+        throw new Error(ERROR.INVALID_ROLE);
+    }
+    if (context.userInfo.role === USER.ROLES.ORG_OWNER && !context?.userInfo?.organization?.orgID) {
+        throw new Error(ERROR.NO_ORG_ASSIGNED);
+    }
+
+    const filters = {};
+    if (context?.userInfo?.role === USER.ROLES.ORG_OWNER) {
+        filters["_id"] = context?.userInfo?.organization?.orgID;
+    }
+
+    return this.listOrganizations(filters);
+  }
+
+  /**
+   * List organizations by an optional set of filters
+   *
+   * @typedef {Object<string, any>} Filters K:V pairs of filters
+   * @param {Filters} [filters] Filters to apply to the query
+   * @returns {Promise<Object[]>} An array of Organizations
+   */
+  async listOrganizations(filters = []) {
     return await this.organizationCollection.aggregate([{ "$match": filters }]);
   }
 }
